Games/Practices of visiting coach
The games and practices are deleted after the event has ended. The team is only deleted if its the last game/practice scheduled.

+++ src/Scheduler/Controllers/HomeController.cs
@@ -160,11 +160,11 @@ public sealed class HomeController : Controller
 			var homeTeam = this.context.Teams.FirstOrDefault(t => t.Id == game.HomeTeamId);
 			var opposingTeam = this.context.Teams.FirstOrDefault(t => t.Id == game.OpposingTeamId);
 			// Delete teams associated with the game that match the user ID
-			if (homeTeam is not null && homeTeam.UserId == userId)
+			if (homeTeam is not null && homeTeam.UserId == userId && gamesToDelete.Count == 1)
 			{
 				this.context.Teams.Remove(homeTeam);
 			}
-			else if (opposingTeam is not null &&  opposingTeam.UserId == userId)
+			else if (opposingTeam is not null &&  opposingTeam.UserId == userId && gamesToDelete.Count == 1)
 			{
 				this.context.Teams.Remove(opposingTeam);
 			}
@@ -180,7 +180,7 @@ public sealed class HomeController : Controller
 		foreach (var practice in practicesToDelete)
 		{
 			// Delete teams associated with the practice that match the user ID
-			if (practice.Team is not null && practice.Team.UserId == userId)
+			if (practice.Team is not null && practice.Team.UserId == userId && practicesToDelete.Count == 1)
 			{
 				this.context.Teams.Remove(practice.Team);
 			}
